+++ product/src/main/java/com/sparta/hotdeal/product/application/dtos/req/product/ReqPostPromotionDto.java
@@ -1,23 +1,19 @@
-package com.sparta.hotdeal.product.application.dtos.product;
+package com.sparta.hotdeal.product.application.dtos.req.product;
 
 import java.sql.Timestamp;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
-@NoArgsConstructor
-@Builder
 @AllArgsConstructor
-public class ResGetPromotionDto {
-    private UUID promotionId;
+@NoArgsConstructor
+@Getter
+public class ReqPostPromotionDto {
     private UUID productId;
     private Timestamp start;
     private Timestamp end;
     private Integer discountRate;
     private Integer discountPrice;
     private Integer quantity;
-    private Integer remaining;
 }
