+++ product/src/main/java/com/sparta/hotdeal/product/application/dtos/req/product/ReqPutPromotionDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotdeal.product.presentation.dtos.product;
+package com.sparta.hotdeal.product.application.dtos.req.product;
 
 import java.sql.Timestamp;
 import java.util.UUID;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class ReqPostPromotionDto {
+public class ReqPutPromotionDto {
     private UUID productId;
     private Timestamp start;
     private Timestamp end;
