+++ product/src/main/java/com/sparta/hotdeal/product/application/dtos/req/review/ReqPutReviewDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotdeal.product.application.dtos.review;
+package com.sparta.hotdeal.product.application.dtos.req.review;
 
 import java.util.List;
 import lombok.Builder;
@@ -6,8 +6,7 @@ import lombok.Getter;
 
 @Getter
 @Builder
-public class ResGetReviewByIdDto {
-    private String nickname;
+public class ReqPutReviewDto {
     private double rating;
     private String review;
     private List<String> images;
