+++ product/src/main/java/com/sparta/hotdeal/product/application/dtos/res/product/ResGetPromotionDto.java
@@ -1,19 +1,23 @@
-package com.sparta.hotdeal.product.presentation.dtos.product;
+package com.sparta.hotdeal.product.application.dtos.res.product;
 
 import java.sql.Timestamp;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@AllArgsConstructor
-@NoArgsConstructor
 @Getter
-public class ReqPutPromotionDto {
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class ResGetPromotionDto {
+    private UUID promotionId;
     private UUID productId;
     private Timestamp start;
     private Timestamp end;
     private Integer discountRate;
     private Integer discountPrice;
     private Integer quantity;
+    private Integer remaining;
 }
