+++ product/src/main/java/com/sparta/hotdeal/product/application/dtos/res/review/ResGetReviewByIdDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotdeal.product.presentation.dtos.review;
+package com.sparta.hotdeal.product.application.dtos.res.review;
 
 import java.util.List;
 import lombok.Builder;
@@ -6,7 +6,8 @@ import lombok.Getter;
 
 @Getter
 @Builder
-public class ReqPutReviewDto {
+public class ResGetReviewByIdDto {
+    private String nickname;
     private double rating;
     private String review;
     private List<String> images;
